requirements.txt
# Core API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
starlette>=0.36.0
websockets>=12.0

# OpenEnv framework (mock for now or valid if in pip)
# openenv

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM client (hackathon requirement)
openai>=1.12.0

# NLP for deterministic graders
spacy>=3.7.4

# HTTP + utilities
httpx>=0.26.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
anyio>=4.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Deployment
huggingface-hub>=0.20.0

# MCP tools (optional)
mcp>=1.0.0
